Move tool classes into their own folder
